beamer: move _ChainIdParam to beamer.util and make it public
It's going to be useful in multiple places.
